if [[ ! -f "$tool_root_path/configure" ]];

+++ pg_dump/build_musl.sh
@@ -17,16 +17,21 @@ echo "::set-output name=tool_version::$tool_version"
 download_url="https://github.com/postgres/postgres/archive/refs/tags/$tool_version.tar.gz"
 echo "::group::prepare sources $download_url"
 
-mkdir -p "$dp0/release/build" && cd "$dp0/release"
-wget "$download_url" -O "tool-$tool_version.tar.gz" && tar -xf "tool-$tool_version.tar.gz"
 tool_root_path="$dp0/release/postgres-$tool_version"
+
+mkdir -p "$dp0/release/build" && cd "$dp0/release"
+
+if [[ ! -f "$tool_root_path/configure" ]]; then
+  wget "$download_url" -O "tool-$tool_version.tar.gz" && tar -xf "tool-$tool_version.tar.gz"
+fi
+
 cd "$tool_root_path"
 
 echo "::endgroup::"
 
 echo "::group::build"
 
-./configure LDFLAGS=-static  --without-readline --without-zlib
+./configure --without-readline --without-zlib
 
 make -j$(nproc)
